Fri, Feb 18, 2022, 2:13 AM +00:00

+++ CS50x/week5/pset5/speller/tester.c
@@ -16,7 +16,7 @@ int main(int argc, char *argv[])
 
     for (int i = 0; i < len; i++)
     {
-        int alph_index = (int) tolower(argv[1][i] % 96);
+        int alph_index = tolower(argv[1][i]) % 96;
         hash_num += (alph_index * pow(26, i));
     }
